Refactor aiohttp webserver code into separate lib file/class.
The webserver/websocket logic is large and complex enough to warrant it's own
file for better readability and isolation. Pull that code out of the main
method and encapsulate it in a new file and class.

With this new class, implement aenter/aexit methods to not only start the
webapp, but also clean it up on exit.