Don't update full_history until after it's checked
When combining cases a bug was introduced which caused the full history to
never be sent to clients on connect. Before checking it to determine if the
full or incremental history needed to be sent to a client, full_history was
always being set to False; wait until after it's checked to update it's value.