Sync all UI state from server
When a user makes a change via the UI (target temps, temp unit) don't locally
apply the change in that client; this creates multiple workflows for updating
the local client making the change vs other clients also monitoring the ibbq.

Instead, just validate and send an event to the server, and include all ibbq
client state in the update messages - all clients will update the UI when they
receive the next update message from the server. This ensures all clients have
a common view of the ibbq state.