Handle the exception when changing settings on disconnected device
When the device is disconnected, setting writes cannot be completed but the
websocket handler wasn't accounting for this. For now, ignore these exceptions
(the client UI will simply not reflect the change for now, no error is shown).

Also, use a more descriptive ConnectionError when writes are attempted on
a disconnected device, rather than RuntimeError.