Reduce unneccessary graph rendering
The graph canvas was being re-rendered every time a websocket event callback
was issued. In cases where there are many message queued up (ex: the browser
throttled/disabled JS because the tab/browser was not in the foreground), when
we start processign all of those messages the UI would hang for sometimes
seconds waiting for a ton of incremental re-renderings.

Instead, be smarter about if/when we re-render the canvas, by adding a wrapper
around graph.render(). The wrapper tracks the last timestamp (ms) since the last
time we tried to render the graph. If it's been tried sufficiently recently,
schedule a timeout to try again a bit later... if there are many such cases this
results in many updates being processed and coalesced into a single re-rendering.

Also, skip rendering the graph if it's not visible; either because a differen tab
on this page is selected, or a the browser tab/window is not active. Hook into both
page and browser state changes to ensure the graph is re-rendered when it becomes
displayed again.