Add (non-persistent) setting to change chart's Y-axis minimum
The chart's Y-axis minimum defaults to 0, but some times it's helpful to
adjust this to better view the data; add a form field in the settings tab
to allow this to be change. This setting is not currently persisted or synced
across clients (unlike temperature unit), it is purely a local client-side
setting.