Change timestamp wire/javascript format to int (ms since epoch)
Rather than converting back and forth between timestamp/string, and using
Date() objects as chart x axis values, pass and store the timestamp as an
integer (ms since epoch) and only convert to a timestamp string when
displaying it on a tooltip. This simplifies a bunch of the handling/computation
and should be a cheaper comparison/lookup in javascript.