Allow downloading chart data and viewing it back
Add the ability to download the chart data to a .json file, and then load those
saved data files back into the rendered chart.

When viewing saved chart data, the websocket to the server is disconnected and
a banner is shown to indicate that the data is not live, along with a button to
re-connect back to the server to view live data again.