Validation and notify file upload parsing failures
Add validations to verify the user selected file is indeed a JSON file in the
expected format/schema. If this validation fails, a toast message is added to
briefly notify the user of an error.

Also switch the banners from custom CSS styles to using bootstrap classes
for coloring/position/display.