Add Cache-Control header to static resource HTTP responses
Add a middleware to add the Cache-Control header to responses when serving up
static resources, with max-age=0. This forces browsers to re-validate any
cached copy of the file they might have.

Without this, browsers are free to cache the file however they want, and chrome
seems to choose to use the file indefinitely without any attempt to check with
the server if the cache is stale. This is particularly annoying on mobile where
there's no easy way to hard-reload the page without cached contents.

With this, chrome re-requests all files with the If-Modified-Since and
If-None-Match headers; in the common case the server just responds with
HTTP 304 (Not Modified) and there's minimal impact to page load time.