Handle repeated connection failures
While the developer console shows connection errors (ex: timeout) as happening
on the line creating the websocket object, they cannot be caught with a
try/catch around the websocket creation as the error happens asynchronously;
thus the try/catch is removed.

Instead, both the onclose() and onerror() callbacks are issued when such a
connection error occurs. The issue is actually a logic bug introduced in
1087efc1 when offlineMode was added. When the onclose() callback is issued,
unless we're in offline mode we need to always create a timeout to call
connectWebsocket() and try to re-connect... not just the first time (i.e. when
the banner isn't showing yet). This ensures we will continue to try to
reconnect the websocket until it succeeds.