Add UI button to power off the server
When run on a raspberry py or other similar machine with no physical power
button, it's annoying to have to SSH into the server to power it off cleanly
when done monitoring temperatures. This change adds a UI button to power off
the server over the web interface.

Since power off is a sensitive operation that some might not want to expose
over the network unauthenticated, this is placed behind a config option that
is disabled by default.