Switch to 0 size markers for chart data points
Setting the marker type to none causes no markers ever; this makes it very hard
to tell where you've clicked and are seeing tooltips on mobile if there's lots
of data.

Instead, leave the marker type as the default circle but set it's size to 0 so
it's not visible by default. This still allows marker to be shown when a data
point is activated by hover/click.