Replace connection dot CSS and text with bootstrap icons
Simplify the UI by removing the connection state colored dot and text
with a monocrome bootstrap icon which changes depending on the state.
This offers a more minimal look and better matches the styleing of the
battery level for a more consistent status bar.

Also clean up the code by factoring out the conneciton state code into
a method.