Make cert rotation handling dynamic based on cert expiration
Parse the certificate expiration timestamp when loading the cert chain, and
while it's expired try reloading the certificates at a much faster interval.
Also check the last modified timestamp of the certificate, and skip reloading
the cert chain if it hasn't changed since the last time we reloaded the certs.